refactor: type
